# index.js
require('dotenv').config();

const express = require('express');
process.on('uncaughtException', (err) => {
  console.error('🚨 Error fatal no atrapado:', err);
});

process.on('unhandledRejection', (reason, promise) => {
  console.error('🚨 Promesa rechazada en el fondo:', reason);
});
const app = express();
const port = 3000;
const swaggerUi = require('swagger-ui-express');
const swaggerDocument = require('./openapi.json');
const { Pool } = require('pg');
const { createClient } = require('@supabase/supabase-js');

const supabaseUrl = process.env.SUPABASE_URL;
const supabaseKey = process.env.SUPABASE_KEY;
const supabase = createClient(supabaseUrl, supabaseKey); // Crear la conexión

const pool = new Pool({
  connectionString: process.env.DATABASE_URL
});

app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

async function initDb() {
  await pool.query(`
    CREATE TABLE IF NOT EXISTS tasks (
      id SERIAL PRIMARY KEY,
      title TEXT NOT NULL,
      done BOOLEAN NOT NULL DEFAULT false
    );
  `);

  console.log("Database initialized");

  const result = await pool.query('SELECT COUNT(*) AS total FROM tasks');
  if (Number(result.rows[0].total) === 0) {
    await pool.query('INSERT INTO tasks (title) VALUES ($1)', ['Comprar leche']);
    await pool.query('INSERT INTO tasks (title) VALUES ($1)', ['Lavar ropa']);
    await pool.query('INSERT INTO tasks (title) VALUES ($1)', ['Cocinar lasaña']);
  }
}

async function main() {
  await initDb();
  app.listen(port, () => {
    console.log(`Server is running on port ${port}`);
  });
}

main();

// Middleware para poder leer JSON en el body de las requests (lo necesitarás en Stage 3)
app.use(express.json());

app.param('id', async (req, res, next, id) => {
  const taskId = Number(id);
  if (isNaN(taskId)) {
    return res.status(400).json({ error: "Invalid ID" });
  }

  try {
    const task = await pool.query('SELECT * FROM tasks WHERE id = $1', [taskId]);
    if (!task.rows[0]) {
      return res.status(404).json({ error: `Task ${id} not found` });
    }

    req.task = task.rows[0];
    req.taskId = taskId;
    next();
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "Internal Server Error" });
  }
});

//== 1 ==

// Endpoint para obtener información de la API
app.get('/', (req, res) => {
  res.json({
    name: "Task API",
    version: "1.0",
    endpoints: ["/tasks"]
  });
});

// Endpoint para obtener el estado de la API
app.get('/health', (req, res) => {
  res.json({ status: "ok" });
});

//== 2 ==

// Endpoint para obtener todas las tareas
app.get('/tasks', async (req, res) => {
  if (req.query.done == "true") {
    const doneTasks = await pool.query('SELECT * FROM tasks WHERE done = true');
    return res.json(doneTasks.rows);
  }
  const result = await pool.query('SELECT * FROM tasks');
  res.json(result.rows);
});

// Endpoint para obtener una tarea por su ID
app.get('/tasks/:id', async (req, res) => {
  res.json(req.task); // usa lo que dejó el app.param
});

//== 3 ==

// Endpoint para crear una nueva tarea
app.post('/tasks', async (req, res) => {
  if (req.body.title == null || req.body.title.trim() === "") {
    return res.status(400).json({ error: "Title is required" });
  }

  try {
    const newTask = await pool.query('INSERT INTO tasks (title) VALUES ($1) RETURNING *', [req.body.title]);
    res.status(201).json(newTask.rows[0]);
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Internal Server Error" });
  }
});


//== 4 ==
// Endpoint para actualizar una tarea existente
app.put('/tasks/:id', async (req, res) => {
  if ((req.body.title == null || req.body.title.trim() === "") && req.body.done == null) {
    return res.status(400).json({ error: "Invalid body" });
  }

  try {
    if (req.body.title != null && req.body.title.trim() !== "") {
      const change = await pool.query('UPDATE tasks SET title = $1 where id = $2', [req.body.title, req.taskId]);
    }
    if (req.body.done != null) {
      const change = await pool.query('UPDATE tasks SET done = $1 where id = $2', [Boolean(req.body.done), req.taskId]);
    }

    const taskUpdated = await pool.query('SELECT * FROM tasks where id = $1', [req.taskId]);
    res.json(taskUpdated.rows[0]);
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Internal Server Error" });
  }
});

// Endpoint para eliminar una tarea existente
app.delete('/tasks/:id', async (req, res) => {
  try {
    const taskDelete = await pool.query('DELETE FROM tasks WHERE id = $1', [req.taskId]);
    res.status(204).end();
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Internal Server Error" });
  }
});

//Endpoint para registrarse
app.post('/auth/signup', async (req, res) => {
  if (req.body.email == null || req.body.email.trim() === "") {
    return res.status(400).json({ error: "Invalid body" });
  }
  if (req.body.password == null || req.body.password.trim() === "") {
    return res.status(400).json({ error: "Invalid body" });
  }
  const { data, error } = await supabase.auth.signUp({
    email: req.body.email,
    password: req.body.password,
  })

  if (error) {
    return res.status(400).json({ error: error.message });
  }

  return res.status(201).json(data.user);
});

// Middleware para la autenticación
async function authMiddleware(req, res, next) {
  const authHeader = req.headers.authorization;

  if (!authHeader || !authHeader.startsWith('Bearer ')) {
    return res.status(401).json({ error: "Access token required" });
  }

  const token = authHeader.slice(7).trim();

  if (token === "") {
    return res.status(401).json({ error: "Access token required" });
  }

  const { data: { user } } = await supabase.auth.getUser(token)

  if (!user) {
    return res.status(401).json({ error: "Invalid or expired token" });
  }

  req.user = user;
  req.token = token;

  next();
}


//Endpoint para loggearse
app.post('/auth/login', async (req, res) => {
  if (req.body.email == null || req.body.email.trim() === "") {
    return res.status(400).json({ error: "Invalid body" });
  }
  if (req.body.password == null || req.body.password.trim() === "") {
    return res.status(400).json({ error: "Invalid body" });
  }

  const { data, error } = await supabase.auth.signInWithPassword({
    email: req.body.email,
    password: req.body.password,
  })

  if (error) {
    return res.status(401).json({ error: "Invalid login credentials" });
  }

  return res.status(200).json({
    "access_token": data.session.access_token,
    "refresh_token": data.session.refresh_token
  });
});

// Endpoint publico inforrmativo
app.get('/public/info', async (req, res) => {
  return res.json({ "message": "Welcome stranger! This info is public." });
});

//Endpoint para verificar el token de acceso
app.get('/protected/profile', authMiddleware, async (req, res) => {
  return res.status(200).json({
    id: req.user.id,
    email: req.user.email,
    created_at: req.user.created_at,
  });
});

app.get('/protected/dashboard', authMiddleware, async (req, res) => {
  return res.status(200).json({ message: `Bienvenido, ${req.user.email}` });
});

// Endpoint cerrar sesión
app.post('/auth/logout', authMiddleware, async (req, res) => {
  try {
    const { error } = await supabase.auth.signOut();
    if (error) {
      return res.status(400).json({ error: error.message });
    }
    return res.status(204).end();
  } catch (err) {
    return res.status(500).json({ error: "Internal server error" });
  }

});